move App file to container folder
