+++ src/containers/Route.container.js
@@ -2,7 +2,7 @@ import React from 'react';
 import { BrowserRouter as Router, Switch, Route } from 'react-router-dom';
 import { Provider } from 'react-redux';
 import store from '../store/index.store';
-import App from '../components/App';
+import App from './App.container';
 import Recipes from '../components/Recipes.component';
 import Recipe from '../components/Recipe.component';
